# HoughCircles rewritten (PR #7434 updated)