### This pullrequest changes

Old pull request #7434 updated:
* merge conflicts fixed
* got rid of outdated cv::Seq containters and SSE2 intrinsics
* performance improved (from 2x to 8x depending on a number of candidate circles)
* radius search partially rewritten (histogram of radiuses is build instead of sorting; it is up to 2x faster for cases with big number of edge points)

### Perf tests results

Time of 8 runs is measured in milliseconds

Canny threshold | Accumulator threshold | Original #detected circles |  New #circles | Original code | Improved Code
--------------------- | ---------------------------- | ---------------------- | ---------------- | ---------------- | --------------------
20  |  10  |  59  |  61  |  3984.81  |  1158.9
20  |  30  |  57  |  54  |  9090.84  |  2812.07
20  |  50  |  52  |  55  |  45642.2  |  6522.96
20  |  100  |  23  |  22  |  282254  |  11317.3
50  |  10  |  58  |  57  |  3057.06  |  754.356
50  |  30  |  58  |  53  |  5116.18  |  1394.3
50  |  50  |  51  |  51  |  19703.4  |  1838.52
50  |  100  |  11  |  11  |  51612.8  |  1582.63
100  |  10  |  57  |  56  |  1890.26  |  524.728
100  |  30  |  54  |  53  |  4605.53  |  681.481
100  |  50  |  34  |  31  |  9136.89  |  830.344
100  |  100  |  5  |  6  |  17737.9  |  617.138
150  |  10  |  54  |  50  |  1339.33  |  434.024
150  |  30  |  47  |  49  |  6773.36  |  468.236
150  |  50  |  13  |  14  |  3191.05  |  379.194
150  |  100  |  4  |  4  |  1517.61  |  263.796


The image was used for perf tests:
![detect_circles_soda](https://user-images.githubusercontent.com/10323511/32551991-7c147c4c-c4a3-11e7-8572-300c63a7172e.jpg)
Several other images showed even worse performance of original code.