@mschoeneck 2017-11-14
:+1: Nice to see that you were inspired by the old pull request and good work in further development